# Ligonine/Ligonine/Ligonine/Data/Models/Department.cs
namespace Ligonine.Data.Models
{
    public class Department
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public string Description { get; set; }

    }
}