# Ligonine/Ligonine/Ligonine/Data/Models/Doctor.cs
namespace Ligonine.Data.Models
{
    public class Doctor
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public int Age { get; set; }
        public string BloodType { get; set; }
        public int DepartmentId { get; set; }

    }
}