# Ligonine/Ligonine/Ligonine/Data/Models/Operation.cs
namespace Ligonine.Data.Models
{
    public class Operation
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public string Description { get; set; }
        public int DoctorId { get; set; }
    }
}